后端-实现题目标签导入导出下载模板接口

+++ springcloud-brushtopic-api/topic-model/src/main/java/com/hao/topic/model/entity/topic/TopicSubject.java
@@ -14,7 +14,7 @@ public class TopicSubject extends BaseEntity {
     private String subjectName;
     private String subjectDesc;
     private String imageUrl;
-    private Integer topicCount;
+    private Long topicCount;
     private String createBy;
     private Long viewCount;
     private Integer status;
